pblio: Reorganized files under spc dir
